Front end I have so far
The about, contact, and view pages are empty right now. The sign up and sign in buttons bring up the forms but do not have functionality yet. My plan for the buttons under categories on the home page is to take the user to the view page and display the cuisine based on which button they clicked.  The most viewed food items are just example images of food.